[sdlf-*] rework sdlf to use nested stacks instead of cfn modules
CloudFormation modules are not available in the Zurich region (eu-central-2)
and this might be the case in other op-in regions too...
(we also cannot rule out the possibility of cfn modules to be 'de-emphasized' one day,
although we have no inside info there)

Cfn modules are used for two purposes in SDLF:
* to package SDLF constructs
* as a UX mechanism, by making SDLF constructs easily deployable as resources in CloudFormation templates

In this commit CloudFormation nested stacks are used instead, relying on some 'clever' use of TemplateURL,
{{resolve:ssm}} and SSM parameter name conventions to keep answering both purposes in a rather similar way.

It's not perfect: seedfarmer is a pain with some symbols ({{}}), and aws-sam-cli doesn't like {{resolve:ssm}}
in TemplateURLs. But it's pretty good... and the following commits should remove the reliance on seedfarmer, as planned.

This commit removes support for cfn modules entirely, this will be restored in a follow-up commit.